Lowercase history table_name before checking if it exists:
We do this because the CREATE TABLE sql command is automatically lower
casing table_name since it is not wrapped in double quotes.  This is an
issue for app labels that have upper case letters.  The table_exists
function is always returning false if the app has upper case letters.
Then, an exception is raised when attempting to create a table with the
same name.